Update Project2 Draft.Rmd
Use the functions in slides to check AUC and show ROC plot, and double density plot also. Still need to improve the function , for now I just use it to check if the features I choose are acceptable. If we use the functions on sildes we would not get a good mark.